requirements.txt
Django==2.2.6
Faker==2.0.3
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
whitenoise==4.1.4
